refactor: sign in layout
