+++ apps/mobile/app/(auth)/_layout.tsx
@@ -1,10 +1,7 @@
 import { StackHeader } from '@/components/navigation';
 import { Stack } from 'expo-router';
-import { useTranslation } from 'react-i18next';
 
 export default function AuthLayout() {
-  const { t } = useTranslation();
-
   return (
     <Stack
       screenOptions={{
@@ -15,7 +12,7 @@ export default function AuthLayout() {
       <Stack.Screen name="guide" options={{ headerShown: false }} />
       <Stack.Screen name="create-kitchen" options={{ title: '' }} />
       <Stack.Screen name="join-kitchen" options={{ title: '' }} />
-      <Stack.Screen name="sign-in" options={{ title: t('auth.signIn.title') }} />
+      <Stack.Screen name="sign-in" options={{ title: '' }} />
     </Stack>
   );
 }
