+++ apps/mobile/app/(auth)/sign-in.tsx
@@ -1,24 +1,27 @@
 import { Formik } from 'formik';
 import { useTranslation } from 'react-i18next';
-import { Alert, StyleSheet } from 'react-native';
+import { StyleSheet } from 'react-native';
 import { router } from 'expo-router';
 import { Button } from 'react-native-paper';
 import * as Yup from 'yup';
 import { SignInDTO } from '@/types';
-import { Surface, ParallaxScrollView, Text, HelloWave } from '@/components/common';
+import { Surface, ParallaxScrollView, Text } from '@/components/common';
 import { PasswordInput, FieldInput } from '@/components/form';
 import { useAuth } from '@/context/AuthContext';
 import { globalStyles } from '@/assets/styles';
+import SignInHero from '@/assets/images/sign-in.svg';
+import { useToast } from '@/context/ToastContext';
 
 export default function SignInScreen() {
+  const { showToast } = useToast();
   const { t } = useTranslation();
   const { signIn } = useAuth();
   const handleSignIn = async (values: SignInDTO) => {
     try {
       await signIn(values.email, values.password);
-      router.push('/');
+      router.replace('/');
     } catch (error) {
-      Alert.alert('Error');
+      showToast('Error');
     }
   };
 
@@ -31,58 +34,62 @@ export default function SignInScreen() {
 
   return (
     <ParallaxScrollView>
-      <Surface style={styles.container}>
-        <Surface style={globalStyles.hero}>
-          <Text className="text-2xl font-bold mb-4">
-            {t('auth.signIn.hero')} <HelloWave />
+      <Surface style={globalStyles.hero}>
+        <SignInHero width="100%" />
+        <Text variant="titleLarge" style={{ textAlign: 'center' }}>
+          {t('auth.signIn.hero')}
+        </Text>
+        <Text variant="labelLarge" style={{ textAlign: 'center', marginTop: 8 }}>
+          {t('auth.signIn.description')}
+        </Text>
+      </Surface>
+
+      <Formik
+        initialValues={{ email: 'admin@test.com', password: 'admin123' }}
+        onSubmit={(values) => {
+          const dto: SignInDTO = {
+            ...values
+          };
+          handleSignIn(dto);
+        }}
+        validationSchema={validationSchema}
+      >
+        {({ handleSubmit, errors }) => (
+          <Surface style={globalStyles.form}>
+            <FieldInput i18nKey="common" name="email" />
+            <PasswordInput i18nKey="common" name="password" />
+
+            <Button
+              style={styles.button}
+              mode="contained"
+              onPress={() => {
+                if (Object.keys(errors).length > 0) {
+                  showToast(Object.values(errors)[0]);
+                  return;
+                }
+                handleSubmit();
+              }}
+            >
+              {t('common.confirm')}
+            </Button>
+          </Surface>
+        )}
+      </Formik>
+
+      <Surface style={{ flex: 1, alignItems: 'center' }}>
+        <Text type="secondary">
+          {t('auth.signIn.haveNoAccount')}
+          <Text type="link" onPress={() => router.push('/(auth)/guide')}>
+            {t('auth.signIn.joinOrCreate')}
           </Text>
-          <Text variant="labelLarge">{t('auth.signIn.description')}</Text>
-        </Surface>
-
-        <Formik
-          initialValues={{ email: 'admin@test.com', password: 'admin123' }}
-          onSubmit={(values) => {
-            const dto: SignInDTO = {
-              ...values
-            };
-            handleSignIn(dto);
-          }}
-          validationSchema={validationSchema}
-        >
-          {({ handleSubmit }) => (
-            <Surface style={globalStyles.form}>
-              <FieldInput i18nKey="common" name="email" />
-
-              <PasswordInput i18nKey="common" name="password" />
-
-              <Button style={styles.button} mode="contained" onPress={() => handleSubmit()}>
-                {t('common.confirm')}
-              </Button>
-            </Surface>
-          )}
-        </Formik>
-
-        <Surface style={{ flex: 1, marginTop: 12, alignItems: 'center' }}>
-          <Text type="secondary">
-            {t('auth.signIn.haveNoAccount')}
-            <Text type="link" onPress={() => router.push('/(auth)/guide')}>
-              {t('auth.signIn.joinOrCreate')}
-            </Text>
-          </Text>
-        </Surface>
+        </Text>
       </Surface>
     </ParallaxScrollView>
   );
 }
 
 const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    width: '100%',
-    gap: 8,
-    justifyContent: 'space-between'
-  },
   button: {
-    marginTop: 32
+    marginTop: 20
   }
 });
