+++ apps/mobile/components/common/index.ts
@@ -10,6 +10,5 @@ export { default as GradientBackground } from './GradientBackground';
 export { default as Modal } from './Modal';
 export * from './List';
 export { default as DigitInput } from './DigitInput';
-export { default as Toast } from './Toast';
 export { default as BottomSheet } from './BottomSheet';
 export { default as BottomSheetTextInput } from './BottomSheetTextInput';
